+++ passkit/lib/src/order/order_line_item.g.dart
@@ -19,16 +19,24 @@ OrderLineItem _$OrderLineItemFromJson(Map<String, dynamic> json) =>
       sku: json['sku'] as String?,
     );
 
-Map<String, dynamic> _$OrderLineItemToJson(OrderLineItem instance) =>
-    <String, dynamic>{
-      'image': instance.image,
-      'price': instance.price,
-      'quantity': instance.quantity,
-      'subtitle': instance.subtitle,
-      'title': instance.title,
-      'gtin': instance.gtin,
-      'sku': instance.sku,
-    };
+Map<String, dynamic> _$OrderLineItemToJson(OrderLineItem instance) {
+  final val = <String, dynamic>{};
+
+  void writeNotNull(String key, dynamic value) {
+    if (value != null) {
+      val[key] = value;
+    }
+  }
+
+  writeNotNull('image', instance.image);
+  writeNotNull('price', instance.price);
+  val['quantity'] = instance.quantity;
+  writeNotNull('subtitle', instance.subtitle);
+  val['title'] = instance.title;
+  writeNotNull('gtin', instance.gtin);
+  writeNotNull('sku', instance.sku);
+  return val;
+}
 
 OrderCurrencyAmount _$OrderCurrencyAmountFromJson(Map<String, dynamic> json) =>
     OrderCurrencyAmount(